rm Serialized types
